+++ junos/resource_security_dynamic_address_feed_server_test.go
@@ -51,6 +51,10 @@ resource "junos_security_dynamic_address_feed_server" "testacc_dyn_add_feed_srv"
     path = "/srx/"
   }
 }
+resource "junos_security_dynamic_address_feed_server" "testacc_dyn_add_feed_srv2" {
+  name = "tfacc_dafeedsr2"
+  url  = "https://example.com/?test=#2"
+}
 `
 }